requirements.txt
pydantic==2.6.2
ip2location-io==1.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
ua-parser==0.18.0
validators==0.22.0
uvicorn==0.27.1
fastapi==0.110.0
APScheduler==3.10.4